+++ NEWS.md
@@ -3,6 +3,8 @@
 * Fixed a bug in woebin_adj function when all_var == FALSE and the breaks of all variables are perfect. 
 * Provide parallel computation (foreach with parallel backend) in the functions of woebin and woebin_ply.
 * Modified scorecard_ply function.
+* Fixed a bug in woebin when there are empty bins based on provided break points. 
+* Added functions ie and ig to calculate information entropy and gini index.
 
 # scorecard 0.1.5